Breaking part3 into multiple folders

+++ part3/eslint.config.mjs
@@ -19,7 +19,7 @@ export default [
     rules: {
       "@stylistic/js/indent": ["error", 2],
       "@stylistic/js/linebreak-style": ["error", "unix"],
-      "@stylistic/js/quotes": ["error", "single"],
+      "@stylistic/js/quotes": ["error", "double"],
       "@stylistic/js/semi": ["error", "never"],
       eqeqeq: "error",
       "no-trailing-spaces": "error",
